refactor: ExpenseForm + IncomeForm 통합 → TransactionForm (#388)
- TransactionForm 공통 컴포넌트 생성 (type prop으로 지출/수입 분기)
- ExpenseForm: 578→11줄 (wrapper)
- IncomeForm: 410→11줄 (wrapper)
- TYPE_CONFIG로 색상/라벨/OCR 유무 관리
- TransactionForm 테스트 14개 추가
- 기존 라우트/import 변경 없음 (wrapper 패턴)
- 전체 894 tests passed
